+++ graal-js/src/com.oracle.truffle.js.test/src/com/oracle/truffle/js/test/interop/resources/defaultclassexporttest.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -38,7 +38,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-import MyMath from 'mymathmodule.js'; // import default class
+import MyMath from './mymathmodule.js'; // import default class
 
 var myMath = new MyMath();